Clarify naming - REST API endpoint publishes function install - doesnt actually install it
